protected update and delete goals routes

+++ backend/controllers/goalController.js
@@ -1,11 +1,12 @@
 const asyncHandler = require('express-async-handler')
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 // desc     Get all goals
 // @route   GET /api/goals
 // access   Private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id })
     res.status(200).json(goals)
 })
 
@@ -18,7 +19,8 @@ const setGoal = asyncHandler(async (req, res) => {
         throw new Error('Please add a text field')
     }
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
     res.status(200).json(goal)
 })
@@ -32,6 +34,21 @@ const updateGoal = asyncHandler(async (req, res) => {
     if(!goal){
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    // check if user exists
+    const user = await User.findById(req.user.id)
+
+    // If no user, raise an error
+    if(!user){
+        res.status(401)
+        throw new Error('User does not exist')
+    }
+
+    // check if user id matches 
+    if(user.id !== goal.user.toString()){
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     // find the goal, update it with request body, create it if it does not exist
@@ -49,6 +66,21 @@ const deleteGoal = asyncHandler(async (req, res) => {
     if(!goal){
         res.status(400)
         throw new Error('Goal does not exist')
+    }
+
+    // check if user exists
+    const user = await User.findById(req.user.id)
+
+    // If no user, raise an error
+    if(!user){
+        res.status(401)
+        throw new Error('User does not exist')
+    }
+
+    // check if user id matches 
+    if(user.id !== goal.user.toString()){
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     await Goal.findByIdAndDelete(req.params.id)
